requirements.txt
streamlit>=1.45
supabase>=2.4
pandas>=2.2
plotly>=5.22
python-dateutil>=2.9
openpyxl>=3.1
python-telegram-bot>=21.0
